refactor: laout as a component
