--- app/assets/stylesheets/posts.scss
@@ -1,3 +1,0 @@
-// Place all the styles related to the posts controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: https://sass-lang.com/
