+++ app/controllers/api/v1/posts_controller.rb
@@ -1,7 +1,7 @@
 class Api::V1::PostsController < ApiController
   before_action :set_post, only: %i[show update destroy]
 
-  rescue_from ActiveRecord::RecordNotFound do |exception|
+  rescue_from ActiveRecord::RecordNotFound do |_exception|
     render json: { error: '404 not found' }, status: 404
   end
 
@@ -38,19 +38,19 @@ class Api::V1::PostsController < ApiController
 
   private
 
-    def set_post
-      @post = Post.find(params[:id])
-    end
+  def set_post
+    @post = Post.find(params[:id])
+  end
 
-    def post_params
-      params.fetch(:post).permit(:title, :body, :post_image)
-    end
+  def post_params
+    params.fetch(:post).permit(:title, :body, :post_image)
+  end
 
-    def render_status_404(exception)
-      render json: { errors: [exception] }, status: 404
-    end
+  def render_status_404(exception)
+    render json: { errors: [exception] }, status: 404
+  end
 
-    def render_status_500(exception)
-      render json: { errors: [exception] }, status: 500
-    end
+  def render_status_500(exception)
+    render json: { errors: [exception] }, status: 500
+  end
 end
