--- app/controllers/posts_controller.rb
@@ -1,53 +1,0 @@
-class PostsController < ApplicationController
-  def index
-    @posts = Post.all
-  end
-
-  def new
-    @post = Post.new
-    @parent = Category.where(ancestry: nil)
-  end
-
-  def create
-    @post = current_user.posts.new(post_params)
-    if @post.save
-      redirect_to posts_path
-    else
-      @post = current_user.posts.new(post_params)
-      render :new
-    end
-  end
-
-  def edit; end
-
-  def update
-    @post = Post.find(params[:id])
-    if @post.update(post_params)
-      redirect_to posts_path
-    else
-      render :edit
-    end
-  end
-
-  def show
-    @post = Post.find(params[:id])
-    @user = @post.user
-    @category_id = @post.category_ids
-    @category = Category.find(@category_id)
-  end
-
-  def destroy
-    @post = Post.find(params[:id])
-    if @post.destroy
-      redirect_to posts_path
-    else
-      render :show
-    end
-  end
-
-  private
-
-  def post_params
-    params.require(:post).permit(:title, :body, :post_image, :category_ids [])
-  end
-end
