+++ app/javascript/components/Like/LikeButton.vue
