+++ app/javascript/components/Post/Modal.vue
