+++ app/javascript/components/Post/PostForm.vue
