+++ app/javascript/packs/index.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import App from '../components/app.vue'
+import App from '../pages/app.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
   const el = document.body.appendChild(document.createElement('hello'))
