+++ app/javascript/pages/posts/PostShow.vue
@@ -18,7 +18,7 @@
 <script>
 import axios from "axios";
 
-import Modal from "./Modal.vue";
+import Modal from "../../components/Post/Modal.vue";
 
 export default {
   components: {
