+++ app/javascript/pages/posts/PostsIndex.vue
