+++ app/views/home/index.html.erb
