--- app/views/layouts/_header.html.erb
@@ -1,27 +1,0 @@
-<nav class="level is-mobile head">
-  <div class="level-left">
-    <p class="level-item">
-      <%= link_to "レシピを書く", new_post_path, class: "navbar-item is-size-7" %>
-    </p>
-  </div>
-  <p class="level-item has-text-centered">
-    <%= link_to "MKgram", posts_path, class: "navbar-item is-size-4", style:"font-family: cursive;" %>
-  </p>
-  <div class="level-right">
-    <p class="level-item navbar-item">
-      <% if user_signed_in? %>
-        <%= link_to user_path(current_user) do %>
-          <span class="icon icon	mx-3">
-            <i class="fa fa-2x fa-user-circle" aria-hidden="true"></i>
-          </span>
-        <% end %>
-      <% else %>
-        <%= link_to sign_in_path do %>
-          <span class="icon icon	mx-3">
-            <i class="fa fa-2x fa-user-circle" aria-hidden="true"></i>
-          </span>
-        <% end %>
-      <% end %>
-    </p>
-  </div>
-</nav>
