--- app/views/posts/edit.html.erb
@@ -1,21 +1,0 @@
-<h2>新規投稿</h2>
-
-<%= form_with model: Post, url: posts_path, do |f| %>
-
-    <div class="field">
-      <%= f.label :post_image, "写真" %><br/>
-      <%= f.file_field :post_image %>
-    </div>
-
-    <div class="field">
-      <%= f.label :title, "タイトル" %><br/>
-      <%= f.text_field :title %>
-    </div>
-
-    <div class="field">
-      <%= f.label :body, "レシピ" %><br/>
-      <%= f.text_area :body %>
-    </div>
-
-<%= f.submit "投稿" %>
-<% end %>
