--- app/views/posts/show.html.erb
@@ -1,11 +1,0 @@
-<%= image_tag @post.post_image.thumb.url %>
-<p><%= @post.title %>by.<%= link_to @user.name, user_path(@user.id)%></p>
-
-<p><%= @post.body %></p>
-<p><%= @category %></p>
-<% if user_signed_in? %>
-  <%= link_to "編集", edit_post_path %>
-  <%= link_to "削除", post_path(@post.id), method: :delete %><br>
-<% end %>
-
-<%= link_to "一覧に戻る", posts_path %>
